requirements.txt
fastapi[standard]==0.115.6
SQLAlchemy==2.0.36
PyJWT==2.10.1
PyMySQL==1.1.1
passlib==1.7.4
aiofiles==24.1.0
python-dotenv==1.0.1

